+++ projects/start-wrt/web/src/app/services/api/live-api.service.ts
@@ -53,6 +53,7 @@ import {
   WanDdnsSetRequest,
   PublishedPortFromApi,
   PublishedPortsSetRequest,
+  AutoForwardFromApi,
   OutboundVpn,
   OutboundVpnCreateRequest,
   OutboundVpnCreateResponse,
@@ -261,6 +262,13 @@ export class LiveApiService extends ApiService {
     return this.rpc.request({ method: 'devices.update', params })
   }
 
+  async devicesSetAutoForward(params: {
+    mac: string
+    allow: boolean
+  }): Promise<null> {
+    return this.rpc.request({ method: 'devices.set-auto-forward', params })
+  }
+
   async devicesForget(params: { mac: string }): Promise<null> {
     return this.rpc.request({ method: 'devices.forget', params })
   }
@@ -333,6 +341,13 @@ export class LiveApiService extends ApiService {
 
   async publishedPortsSet(params: PublishedPortsSetRequest): Promise<null> {
     return this.rpc.request({ method: 'published-ports.set', params })
+  }
+
+  async publishedPortsAutoList(): Promise<AutoForwardFromApi[]> {
+    return this.rpc.request({
+      method: 'published-ports.auto-list',
+      params: {},
+    })
   }
 
   async vpnClientList(): Promise<OutboundVpn[]> {